+++ docs/changelog.rst
@@ -23,6 +23,8 @@ Unreleased
 * :gh:issue:`987` Support Python 3.11
 * :gh:issue:`885` Fix :py:exc:`PermissionError` in :py:mod:`importlib` when
   becoming an unprivileged user with Python 3.x
+* :gh:issue:`1033` Support `PEP 451 <https://peps.python.org/pep-0451/>,
+  required by Python 3.12
 
 
 v0.3.4 (2023-07-02)